+++ app/src/androidTest/java/com/github/se/gatherspot/ui/CreateEventTest.kt
@@ -21,7 +21,7 @@ import com.github.se.gatherspot.model.Interests
 import com.github.se.gatherspot.model.location.Location
 import com.github.se.gatherspot.screens.EventDataFormScreen
 import com.github.se.gatherspot.ui.TopLevelDestinations.EventsViewModel
-import com.github.se.gatherspot.ui.event.CreateEvent
+import com.github.se.gatherspot.ui.eventUI.CreateEvent
 import com.github.se.gatherspot.ui.navigation.NavigationActions
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import org.junit.After
