+++ app/src/androidTest/java/com/github/se/gatherspot/ui/EditEventTest.kt
@@ -11,7 +11,7 @@ import com.github.se.gatherspot.model.event.EventStatus
 import com.github.se.gatherspot.model.location.Location
 import com.github.se.gatherspot.screens.EventDataFormScreen
 import com.github.se.gatherspot.ui.TopLevelDestinations.EventsViewModel
-import com.github.se.gatherspot.ui.event.EditEvent
+import com.github.se.gatherspot.ui.eventUI.EditEvent
 import com.github.se.gatherspot.ui.navigation.NavigationActions
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import java.time.LocalDate
