+++ app/src/androidTest/java/com/github/se/gatherspot/ui/EventUITest.kt
@@ -18,8 +18,8 @@ import com.github.se.gatherspot.model.Profile
 import com.github.se.gatherspot.model.event.Event
 import com.github.se.gatherspot.screens.EventUIScreen
 import com.github.se.gatherspot.ui.TopLevelDestinations.EventsViewModel
-import com.github.se.gatherspot.ui.event.EventUI
-import com.github.se.gatherspot.ui.event.EventUIViewModel
+import com.github.se.gatherspot.ui.eventUI.EventUI
+import com.github.se.gatherspot.ui.eventUI.EventUIViewModel
 import com.github.se.gatherspot.ui.navigation.NavigationActions
 import com.google.firebase.auth.FirebaseAuth
 import io.github.kakaocup.compose.node.element.ComposeScreen
