+++ app/src/main/java/com/github/se/gatherspot/MainActivity.kt
@@ -27,16 +27,16 @@ import com.github.se.gatherspot.model.MapViewModel
 import com.github.se.gatherspot.model.chat.ChatViewModel
 import com.github.se.gatherspot.model.chat.ChatsListViewModel
 import com.github.se.gatherspot.model.event.Event
-import com.github.se.gatherspot.ui.event.EventUIViewModel
+import com.github.se.gatherspot.ui.eventUI.EventUIViewModel
 import com.github.se.gatherspot.model.utils.LocalDateDeserializer
 import com.github.se.gatherspot.model.utils.LocalDateSerializer
 import com.github.se.gatherspot.model.utils.LocalTimeDeserializer
 import com.github.se.gatherspot.model.utils.LocalTimeSerializer
 import com.github.se.gatherspot.ui.TopLevelDestinations.ChatUI
 import com.github.se.gatherspot.ui.TopLevelDestinations.Chats
-import com.github.se.gatherspot.ui.event.CreateEvent
-import com.github.se.gatherspot.ui.event.EditEvent
-import com.github.se.gatherspot.ui.event.EventUI
+import com.github.se.gatherspot.ui.eventUI.CreateEvent
+import com.github.se.gatherspot.ui.eventUI.EditEvent
+import com.github.se.gatherspot.ui.eventUI.EventUI
 import com.github.se.gatherspot.ui.TopLevelDestinations.Events
 import com.github.se.gatherspot.ui.TopLevelDestinations.LogIn
 import com.github.se.gatherspot.ui.TopLevelDestinations.Map
