+++ app/src/main/java/com/github/se/gatherspot/model/EventUtils.kt
@@ -11,7 +11,7 @@ import com.github.se.gatherspot.model.event.DraftEvent
 import com.github.se.gatherspot.model.event.Event
 import com.github.se.gatherspot.model.event.EventStatus
 import com.github.se.gatherspot.model.location.Location
-import com.github.se.gatherspot.ui.event.EventAction
+import com.github.se.gatherspot.ui.eventUI.EventAction
 import java.io.IOException
 import java.time.LocalDate
 import java.time.LocalTime
