+++ app/src/main/java/com/github/se/gatherspot/ui/eventUI/CreateEvent.kt
@@ -1,4 +1,4 @@
-package com.github.se.gatherspot.ui.event
+package com.github.se.gatherspot.ui.eventUI
 
 // GUI to create an event
