+++ app/src/main/java/com/github/se/gatherspot/ui/eventUI/EditEvent.kt
@@ -1,4 +1,4 @@
-package com.github.se.gatherspot.ui.event
+package com.github.se.gatherspot.ui.eventUI
 
 import androidx.compose.runtime.Composable
 import com.github.se.gatherspot.model.EventUtils
